+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Linter's execution action can be set by the `phpCodeSniffer.lintAction` option.
 
 ## [1.3.0] - 2021-05-24
 ### Fixed
@@ -22,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.1.0] - 2021-03-04
 ### Added
-- Ignore diagnostics for files using `phpCodeSniffer.ignorePatterns` option.
+- Ignore diagnostics for files using the `phpCodeSniffer.ignorePatterns` option.
 
 ### Fixed
 - "Ignore * for this line" action should be present for all diagnostics.